cache: outsource responsibilities to multiple files
