+++ packages/cache/src/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_TRIM = 600;
+export const DEFAULT_TTL = 3600;
