+++ packages/cache/src/util.ts
@@ -1,0 +1,3 @@
+export function nowInSeconds() {
+  return new Date().getTime() / 1000;
+}
